type update

+++ sinn/graph/graph_pyg.py
@@ -78,7 +78,7 @@ def create_labeled_supercell(data: torch.Tensor, n: int, lattice: torch.Tensor =
     norm_props = torch.ceil(lattice_vector_norm_prop * n).int()
 
     # create the atom ids
-    atom_id: torch.IntTensor = torch.arange(n_atoms, dtype=torch.int, device=device)
+    atom_id: torch.IntTensor = torch.arange(n_atoms, dtype=torch.int, device=device).int()
     atom_id: torch.IntTensor = atom_id.repeat(n**3)
 
     # create the cell ids
